GestInvLab_Backend/requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.3
Django==5.2.8
djangorestframework==3.16.1
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.2.1
sqlparse==0.5.3
